+++ testutils/logger.go
@@ -21,13 +21,52 @@
 package testutils
 
 import (
+	"bytes"
 	"fmt"
 	"strings"
-	"sync/atomic"
+	"sync"
 	"testing"
+	"time"
 
 	"github.com/uber/tchannel-go"
+	"github.com/uber/tchannel-go/atomic"
 )
+
+// writer is shared between multiple loggers, and serializes acccesses to
+// the underlying buffer.
+type writer struct {
+	sync.Mutex
+	buf *bytes.Buffer
+}
+
+// testLogger is a logger that writes all output to a buffer, and can report
+// the logs if the test has failed.
+type testLogger struct {
+	t      testing.TB
+	fields tchannel.LogFields
+	w      *writer
+}
+
+type errorLoggerState struct {
+	matchCount []atomic.Uint32
+}
+
+type errorLogger struct {
+	tchannel.Logger
+	t testing.TB
+	v *LogVerification
+	s *errorLoggerState
+}
+
+func newWriter() *writer {
+	return &writer{buf: &bytes.Buffer{}}
+}
+
+func (w *writer) withLock(f func(*bytes.Buffer)) {
+	w.Lock()
+	f(w.buf)
+	w.Unlock()
+}
 
 // Matches returns true if the message and fields match the filter.
 func (f LogFilter) Matches(msg string, fields tchannel.LogFields) bool {
@@ -59,16 +98,67 @@ func (f LogFilter) Matches(msg string, fields tchannel.LogFields) bool {
 
 	return true
 }
-
-type errorLoggerState struct {
-	matchCount []uint32
+func newTestLogger(t testing.TB) testLogger {
+	return testLogger{t, nil, newWriter()}
 }
 
-type errorLogger struct {
-	tchannel.Logger
-	t testing.TB
-	v *LogVerification
-	s *errorLoggerState
+func (l testLogger) Enabled(level tchannel.LogLevel) bool {
+	return true
+}
+
+func (l testLogger) log(prefix string, msg string) {
+	logLine := fmt.Sprintf("%s [%v] %v %v\n", time.Now().Format("15:04:05.000000"), prefix, msg, l.Fields())
+	l.w.withLock(func(w *bytes.Buffer) {
+		w.WriteString(logLine)
+	})
+}
+
+func (l testLogger) Fatal(msg string) {
+	l.log("F", msg)
+}
+
+func (l testLogger) Error(msg string) {
+	l.log("E", msg)
+}
+
+func (l testLogger) Warn(msg string) {
+	l.log("W", msg)
+}
+
+func (l testLogger) Info(msg string) {
+	l.log("I", msg)
+}
+
+func (l testLogger) Infof(msg string, args ...interface{}) {
+	l.log("I", fmt.Sprintf(msg, args...))
+}
+
+func (l testLogger) Debug(msg string) {
+	l.log("D", msg)
+}
+
+func (l testLogger) Debugf(msg string, args ...interface{}) {
+	l.log("D", fmt.Sprintf(msg, args...))
+}
+
+func (l testLogger) Fields() tchannel.LogFields {
+	return l.fields
+}
+
+func (l testLogger) WithFields(fields ...tchannel.LogField) tchannel.Logger {
+	existing := len(l.Fields())
+	newFields := make(tchannel.LogFields, existing+len(fields))
+	copy(newFields, l.Fields())
+	copy(newFields[existing:], fields)
+	return testLogger{l.t, newFields, l.w}
+}
+
+func (l testLogger) report() {
+	if l.t.Failed() {
+		l.w.withLock(func(w *bytes.Buffer) {
+			l.t.Logf("Debug logs:\n%s", w.String())
+		})
+	}
 }
 
 // checkFilters returns whether the message can be ignored by the filters.
@@ -84,7 +174,7 @@ func (l errorLogger) checkFilters(msg string) bool {
 		return false
 	}
 
-	matchCount := atomic.AddUint32(&l.s.matchCount[match], 1)
+	matchCount := l.s.matchCount[match].Inc()
 	return uint(matchCount) <= l.v.Filters[match].Count
 }
 func (l errorLogger) checkErr(prefix, msg string) {